Add 'cloud' option docs; remove 'enhanced' backend distinction
These changes add documentation for the new `cloud` option for Terraform
Cloud added in Terraform 1.1.

A big reason for the native integration is to remove the confusing
vestige of 'enhanced' backends that never came to be, beyond the
'remote' backend for Terraform Enterprise specifically. To that end,
these changes also remove the 'type' ('standard' vs. 'enhanced') of
backends from the documentation, leaving the term 'backend' to solely be
used to mean a plugin for state snapshot storage (with the remote
backend being the single exception).